backend/requirements.txt
flask
flask-cors
transformers
torch
huggingface_hub
python-dotenv